ssa: avoid potential cyclic import
This moves things into separate packages to avoid a potential cyclic
import as soon as we would like to utilize `jsondiff` in `ssa` itself.
