+++ ssa/utils/object.go
@@ -1,0 +1,124 @@
+/*
+Copyright 2023 The Flux authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package utils
+
+import (
+	"encoding/json"
+	"io"
+	"strings"
+
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime"
+	yamlutil "k8s.io/apimachinery/pkg/util/yaml"
+	"sigs.k8s.io/yaml"
+)
+
+// ReadObject decodes a YAML or JSON document from the given reader into an unstructured Kubernetes API object.
+func ReadObject(r io.Reader) (*unstructured.Unstructured, error) {
+	reader := yamlutil.NewYAMLOrJSONDecoder(r, 2048)
+	obj := &unstructured.Unstructured{}
+	err := reader.Decode(obj)
+	if err != nil {
+		return nil, err
+	}
+
+	return obj, nil
+}
+
+// ReadObjects decodes the YAML or JSON documents from the given reader into unstructured Kubernetes API objects.
+// The documents which do not subscribe to the Kubernetes Object interface, are silently dropped from the result.
+func ReadObjects(r io.Reader) ([]*unstructured.Unstructured, error) {
+	reader := yamlutil.NewYAMLOrJSONDecoder(r, 2048)
+	objects := make([]*unstructured.Unstructured, 0)
+
+	for {
+		obj := &unstructured.Unstructured{}
+		err := reader.Decode(obj)
+		if err != nil {
+			if err == io.EOF {
+				err = nil
+				break
+			}
+			return objects, err
+		}
+
+		if obj.IsList() {
+			err = obj.EachListItem(func(item runtime.Object) error {
+				obj := item.(*unstructured.Unstructured)
+				objects = append(objects, obj)
+				return nil
+			})
+			if err != nil {
+				return objects, err
+			}
+			continue
+		}
+
+		if IsKubernetesObject(obj) && !IsKustomization(obj) {
+			objects = append(objects, obj)
+		}
+	}
+
+	return objects, nil
+}
+
+// ObjectToYAML encodes the given Kubernetes API object to YAML.
+func ObjectToYAML(object *unstructured.Unstructured) string {
+	var builder strings.Builder
+	data, err := yaml.Marshal(object)
+	if err != nil {
+		return ""
+	}
+	builder.Write(data)
+	builder.WriteString("---\n")
+
+	return builder.String()
+}
+
+// ObjectsToYAML encodes the given Kubernetes API objects to a YAML multi-doc.
+func ObjectsToYAML(objects []*unstructured.Unstructured) (string, error) {
+	var builder strings.Builder
+	for _, obj := range objects {
+		data, err := yaml.Marshal(obj)
+		if err != nil {
+			return "", err
+		}
+		builder.Write(data)
+		builder.WriteString("---\n")
+	}
+	return builder.String(), nil
+}
+
+// ObjectsToJSON encodes the given Kubernetes API objects to a YAML multi-doc.
+func ObjectsToJSON(objects []*unstructured.Unstructured) (string, error) {
+	list := struct {
+		ApiVersion string                       `json:"apiVersion,omitempty"`
+		Kind       string                       `json:"kind,omitempty"`
+		Items      []*unstructured.Unstructured `json:"items,omitempty"`
+	}{
+		ApiVersion: "v1",
+		Kind:       "List",
+		Items:      objects,
+	}
+
+	data, err := json.MarshalIndent(list, "", "    ")
+	if err != nil {
+		return "", err
+	}
+
+	return string(data), nil
+}
